backend/requirements.txt
fastapi
uvicorn
bcrypt
python-jose[cryptography]
databases
sqlalchemy
pydantic
pytest
httpx